@@ -12,35 +12,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.11",
-    "@angular/common": "~12.2.11",
-    "@angular/compiler": "~12.2.11",
-    "@angular/core": "~12.2.16",
-    "@angular/forms": "~12.2.11",
-    "@angular/localize": "~12.2.11",
-    "@angular/platform-browser": "~12.2.11",
-    "@angular/platform-browser-dynamic": "~12.2.11",
-    "@angular/router": "~12.2.11",
+    "@angular/animations": "^14.2.1",
+    "@angular/common": "^14.2.1",
+    "@angular/compiler": "^14.2.1",
+    "@angular/core": "^14.2.1",
+    "@angular/forms": "^14.2.1",
+    "@angular/localize": "^14.2.1",
+    "@angular/platform-browser": "^14.2.1",
+    "@angular/platform-browser-dynamic": "^14.2.1",
+    "@angular/router": "^14.2.1",
     "@digitalspace/bcparks-bootstrap-theme": "^1.0.6",
-    "@ng-bootstrap/ng-bootstrap": "10.0.0",
+    "@ng-bootstrap/ng-bootstrap": "^13.0.0",
     "guid-typescript": "^1.0.9",
     "luxon": "^2.4.0",
     "material-icons": "^0.6.3",
     "ngx-pagination": "^5.1.0",
     "ngx-toastr": "^13.2.1",
     "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.17",
+    "@angular-devkit/build-angular": "^14.2.2",
     "@angular-eslint/builder": "12.7.0",
     "@angular-eslint/eslint-plugin": "12.7.0",
     "@angular-eslint/eslint-plugin-template": "12.7.0",
-    "@angular-eslint/schematics": "12.7.0",
     "@angular-eslint/template-parser": "12.7.0",
-    "@angular/cli": "~12.2.17",
-    "@angular/compiler-cli": "~12.2.11",
+    "@angular/cli": "^14.2.2",
+    "@angular/compiler-cli": "^14.2.1",
     "@types/jasmine": "~3.6.0",
     "@types/node": "^12.11.1",
     "@typescript-eslint/eslint-plugin": "4.28.2",
@@ -54,6 +52,7 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ts-node": "~8.3.0",
-    "typescript": "~4.3.5"
+    "tslib": "^2.4.0",
+    "typescript": "~4.6.4"
   }
 }
